+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to pylinkage are documented here.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- ``examples/strider.py`` from [leggedsnake](https://github.com/HugoFara/leggedsnake), 
+based on the [Strider Linkage](https://www.diywalkers.com/strider-linkage-plans.html)
+
 ## [0.5.3] - 2023-06-23
 
 ### Added in 0.5.3